refactor(components): extract notifications popover
- Add components/notifications/NotificationPopover.vue to encapsulate
  notification fetching, infinite scroll and UI
- Move NotificationItem.vue into components/notifications/
- Replace defineModel('isOpen') in GeneralSettingsDialog with a local
  ref
