+++ frontend/package.json
@@ -56,5 +56,5 @@
     "vitest": "^3.2.4",
     "vue-tsc": "^2.2.10"
   },
-  "packageManager": "pnpm@10.15.1"
+  "packageManager": "pnpm@10.33.0"
 }
