+++ frontend/src/components/notifications/NotificationItem.vue
